requirements.txt
emails==0.6
python-dotenv==0.9.1